actor fix

+++ agent/core/src/main/java/com/agentengine/agent/core/session/state/RolledBackRun.java
@@ -1,3 +1,3 @@
 package com.agentengine.agent.core.session.state;
 
-public record RolledBackRun(String runId, long sequence) {}
+public record RolledBackRun(String runId, int sequence) {}
